* ProgramFOX/30secs:
  i += 1  ->  i++
  Implemented 3/2 and 3/4 time controls
  Improve sliderTime
  Fix increment and correspondence slider
  Better way to check if $input is time slider
  Update modules/chess
  Implemented 30second-games

+++ modules/setup/src/main/FriendConfig.scala
@@ -8,7 +8,7 @@ import lila.rating.RatingRange
 case class FriendConfig(
     variant: chess.variant.Variant,
     timeMode: TimeMode,
-    time: Int,
+    time: Double,
     increment: Int,
     days: Int,
     mode: Mode,
@@ -37,7 +37,7 @@ case class FriendConfig(
 
 object FriendConfig extends BaseHumanConfig {
 
-  def <<(v: Int, tm: Int, t: Int, i: Int, d: Int, m: Option[Int], c: String, fen: Option[String]) =
+  def <<(v: Int, tm: Int, t: Double, i: Int, d: Int, m: Option[Int], c: String, fen: Option[String]) =
     new FriendConfig(
       variant = chess.variant.Variant(v) err "Invalid game variant " + v,
       timeMode = TimeMode(tm) err s"Invalid time mode $tm",
@@ -51,7 +51,7 @@ object FriendConfig extends BaseHumanConfig {
   val default = FriendConfig(
     variant = variantDefault,
     timeMode = TimeMode.Unlimited,
-    time = 5,
+    time = 5d,
     increment = 8,
     days = 2,
     mode = Mode.default,
@@ -65,7 +65,7 @@ object FriendConfig extends BaseHumanConfig {
     def reads(r: BSON.Reader): FriendConfig = FriendConfig(
       variant = chess.variant.Variant orDefault (r int "v"),
       timeMode = TimeMode orDefault (r int "tm"),
-      time = r int "t",
+      time = r double "t",
       increment = r int "i",
       days = r int "d",
       mode = Mode orDefault (r int "m"),
